+++ rest_email_auth/app_settings.py
@@ -5,6 +5,8 @@ The setting implementation is modeled off of "Django Allauth's" from:
 https://github.com/pennersr/django-allauth/blob/master/allauth/account/app_settings.py
 """
 
+import sys
+
 
 class AppSettings(object):
 
@@ -13,8 +15,9 @@ class AppSettings(object):
         Perform some basic settings checks.
         """
         # The user must provide a URL template for the verification
-        # endpoint.
+        # and password reset endpoints.
         assert self.EMAIL_VERIFICATION_URL
+        assert self.PASSWORD_RESET_URL
 
     def _setting(self, name, default):
         """
@@ -72,11 +75,41 @@ class AppSettings(object):
         """
         return self._setting('EMAIL_VERIFICATION_URL', '')
 
+    @property
+    def PASSWORD_RESET_EXPIRATION(self):
+        """
+        The duration that a password reset token is valid for.
+
+        Defaults to 1 hour.
+        """
+        import datetime
+
+        return self._setting(
+            'PASSWORD_RESET_EXPIRATION',
+            datetime.timedelta(hours=1))
+
+    @property
+    def PASSWORD_RESET_URL(self):
+        """
+        The template to use for the password reset url.
+        """
+        return self._setting('PASSWORD_RESET_URL', '')
+
+    @property
+    def REGISTRATION_SERIALIZER(self):
+        """
+        The serializer class used for registering new users.
+        """
+        from django.utils.module_loading import import_string
+
+        return import_string(
+            self._setting(
+                'REGISTRATION_SERIALIZER',
+                'rest_email_auth.serializers.RegistrationSerializer'))
+
 
 # Ugly? Guido recommends this himself ...
 # http://mail.python.org/pipermail/python-ideas/2012-May/014969.html
-import sys      # noqa
-
 
 app_settings = AppSettings()
 app_settings.__name__ = __name__