+++ rest_email_auth/serializers.py
@@ -17,9 +17,6 @@ from rest_email_auth import models
 logger = logging.getLogger(__name__)
 
 
-UserModel = get_user_model()
-
-
 class EmailSerializer(serializers.ModelSerializer):
     """
     Serializer for email addresses.
@@ -199,6 +196,99 @@ class EmailVerificationSerializer(serializers.Serializer):
         return key
 
 
+class PasswordResetRequestSerializer(serializers.Serializer):
+    """
+    Serializer for requesting a password reset.
+    """
+    email = serializers.EmailField(
+        help_text=_("The email address to send the password reset to."))
+
+    def save(self):
+        """
+        Send out a password reset if the provided data is valid.
+
+        If the provided email address exists and is verified, a reset
+        email is sent to the address.
+
+        Returns:
+            The password reset token if it was returned and ``None``
+            otherwise.
+        """
+        try:
+            email = models.EmailAddress.objects.get(
+                email=self.validated_data['email'],
+                is_verified=True)
+        except models.EmailAddress.DoesNotExist:
+            return None
+
+        token = models.PasswordResetToken.objects.create(email=email)
+        token.send()
+
+        return token
+
+
+class PasswordResetSerializer(serializers.Serializer):
+    """
+    Serializer for reseting a user's password.
+    """
+    key = serializers.UUIDField(
+        help_text=_('The key received by the user in the password reset '
+                    'email.'),
+        write_only=True)
+    password = serializers.CharField(
+        help_text=_("The user's new password."),
+        style={'input_type': 'password'},
+        write_only=True)
+
+    def save(self):
+        """
+        Reset the user's password if the provided information is valid.
+        """
+        token = models.PasswordResetToken.objects.get(
+            key=self.validated_data['key'])
+
+        token.email.user.set_password(self.validated_data['password'])
+        token.email.user.save()
+
+        logger.info("Reset password for %s", token.email.user)
+
+        token.delete()
+
+    def validate_key(self, key):
+        """
+        Validate the provided reset key.
+
+        Returns:
+            The validated key.
+
+        Raises:
+            serializers.ValidationError:
+                If the provided key does not exist.
+        """
+        if not models.PasswordResetToken.valid_tokens.filter(key=key).exists():
+            raise serializers.ValidationError(
+                _("The provided reset token does not exist, or is expired."))
+
+        return key
+
+    def validate_password(self, password):
+        """
+        Validate the provided password by running it through Django's
+        password validation system.
+
+        Returns:
+            The validated password.
+
+        Raises:
+            ValidationError:
+                If the provided password does not pass the configured
+                password validators.
+        """
+        password_validation.validate_password(password)
+
+        return password
+
+
 class RegistrationSerializer(serializers.ModelSerializer):
     """
     Serializer for registering new users.
@@ -212,8 +302,8 @@ class RegistrationSerializer(serializers.ModelSerializer):
                 'write_only': True,
             },
         }
-        fields = (UserModel.USERNAME_FIELD, 'email', 'password')
-        model = UserModel
+        fields = (get_user_model().USERNAME_FIELD, 'email', 'password')
+        model = get_user_model()
 
     def create(self, validated_data):
         """
@@ -236,7 +326,9 @@ class RegistrationSerializer(serializers.ModelSerializer):
         email = validated_data.pop('email')
         password = validated_data.pop('password')
 
-        user = UserModel(**validated_data)
+        # We don't save the user instance yet in case the provided email
+        # address already exists.
+        user = get_user_model()(**validated_data)
         user.set_password(password)
 
         # We set an ephemeral email property so that it is included in