+++ docs/CHANGELOG.rst
@@ -2,6 +2,20 @@ Changelog
 =========
 
 
+v0.4.0
+------
+
+Features
+  * :issue:`30`: Add endpoints to request/perform a password reset.
+  * :issue:`37`: Allow a custom registration serializer to be provided.
+
+Documentation
+  * :issue:`29`: Fix typo with installation instructions.
+
+Miscellaneous
+  * :issue:`33`: Fix issue with deployment process breaking example project requirements.
+
+
 v0.3.1
 ------